Remove last `impl_stubs!` bits and move some of them into wasm executor for tests
