+++ core/sr-io/without_std.rs
@@ -285,60 +285,3 @@ impl Printable for u64 {
 pub fn print<T: Printable + Sized>(value: T) {
 	value.print();
 }
-
-#[macro_export]
-macro_rules! impl_stubs {
-	( $( $new_name:ident $($nodecode:ident)* => $invoke:expr ),* ) => {
-		$(
-			impl_stubs!(@METHOD $new_name $($nodecode)* => $invoke);
-		)*
-	};
-	( @METHOD $new_name:ident NO_DECODE => $invoke:expr ) => {
-		#[no_mangle]
-		pub fn $new_name(input_data: *mut u8, input_len: usize) -> u64 {
-			let input: &[u8] = if input_len == 0 {
-				&[0u8; 0]
-			} else {
-				unsafe {
-					$crate::slice::from_raw_parts(input_data, input_len)
-				}
-			};
-
-			let output: $crate::rstd::vec::Vec<u8> = $invoke(input);
-			let res = output.as_ptr() as u64 + ((output.len() as u64) << 32);
-
-			// Leak the output vector to avoid it being freed.
-			// This is fine in a WASM context since the heap
-			// will be discarded after the call.
-			::core::mem::forget(output);
-			res
-		}
-	};
-	( @METHOD $new_name:ident => $invoke:expr ) => {
-		#[no_mangle]
-		pub fn $new_name(input_data: *mut u8, input_len: usize) -> u64 {
-			let mut input = if input_len == 0 {
-				&[0u8; 0]
-			} else {
-				unsafe {
-					$crate::slice::from_raw_parts(input_data, input_len)
-				}
-			};
-
-			let input = match $crate::codec::Decode::decode(&mut input) {
-				Some(input) => input,
-				None => panic!("Bad input data provided to {}", stringify!($name)),
-			};
-
-			let output = ($invoke)(input);
-			let output = $crate::codec::Encode::encode(&output);
-			let res = output.as_ptr() as u64 + ((output.len() as u64) << 32);
-
-			// Leak the output vector to avoid it being freed.
-			// This is fine in a WASM context since the heap
-			// will be discarded after the call.
-			::core::mem::forget(output);
-			res
-		}
-	}
-}
